Rename StateMachine to StateContext
